pyproject.toml
[tool.poetry]
name = "surveyhub"
version = "0.1.0"
description = "SurveyHub Django application"
authors = ["Bharat Kumar <kumar.bhart28@gmail.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
Django = ">=5.0"
setuptools = ">=68.0"
djangorestframework = ">=3.16.0"
django-filter = ">=25.1,<26.0"
pillow = ">=11.3.0,<12.0.0"
pandas = ">=2.3.0,<3.0.0"
numpy = ">=2.3.1,<3.0.0"
matplotlib = ">=3.10.3,<4.0.0"
seaborn = ">=0.13.2,<0.14.0"
openpyxl = ">=3.1.5,<4.0.0"
reportlab = ">=4.4.2,<5.0.0"
channels = ">=4.2.2,<5.0.0"
channels-redis = ">=4.2.1,<5.0.0"
django-extensions = ">=4.1,<5.0"
python-decouple = "==3.8"
psycopg2-binary = ">=2.9.10,<3.0.0"
gunicorn = ">=23.0.0,<24.0.0"
whitenoise = ">=6.9.0,<7.0.0"
django-debug-toolbar = ">=5.2.0,<6.0.0"
celery = ">=5.5.3"
djangorestframework-simplejwt = ">=5.5.0,<6.0.0"
django-cors-headers = ">=4.7.0"
jinja2 = ">=3.1.6,<4.0.0"
requests = ">=2.32.4,<3.0.0"
asgiref = ">=3.6.0"
inflection = "==0.5.1"
packaging = ">=23.0"
factory-boy = ">=3.3.3,<4.0.0"
pytest-django = ">=4.11.1,<5.0.0"
pyjwt = "<2.10.0"
urllib3 = ">=2.5.0,<3.0.0"
charset-normalizer = ">=3.4.2,<4.0.0"
drf-spectacular = ">=0.28.0,<0.29.0"
django-redis = "^6.0.0"
drf-nested-routers = "^0.94.2"

[tool.poetry.group.dev.dependencies]
pytest-django = ">=4.11.1,<5.0.0"
factory-boy = ">=3.3.3,<4.0.0"
black = "25.1.0"
isort = ">=5.13"
pre-commit = ">=3.7"
ruff = "0.12.1"
coverage = ">=7.5"
pytest-cov = ">=5.0"
codecov = ">=2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [".venv", "migrations", "settings"]

[tool.ruff.lint]
extend-select = ["F", "E", "W", "I", "UP", "N"]
ignore = ["E501", "E402", "F403", "F405"]
